test(models): consolidate coverage tests
